@@ -1,39 +1,41 @@
 {
   "name": "stardust",
   "displayName": "StarDust",
-  "description": "Snippets for Terraria mod developers who prefer to use Visual Studio Code or, for some reason, cannot use Visual Studio Community.",
-  "version": "1.0.0",
-  "icon": "images/icon.png",
-  "repository": "https://github.com/Excali-bee/Stardust",
-  "publisher": "VitorFazoli",
+  "description": "Snippets for Terraria mod developers to make their lives easier.",
+  "version": "1.0.1",
   "engines": {
-    "vscode": "^1.91.0"
+    "vscode": "^1.93.0"
   },
   "categories": [
     "Snippets"
   ],
   "activationEvents": [],
-  "main": "./extension.js",
+  "main": "./out/extension.js",
   "contributes": {
     "snippets": [
       {
         "language": "csharp",
-        "path": "./snippets/csharp.json"
+        "path": "snippets/csharp.json"
       }
     ]
   },
   "scripts": {
-    "lint": "eslint .",
-    "pretest": "npm run lint",
+    "vscode:prepublish": "npm run compile",
+    "compile": "tsc -p ./",
+    "watch": "tsc -watch -p ./",
+    "pretest": "npm run compile && npm run lint",
+    "lint": "eslint src",
     "test": "vscode-test"
   },
   "devDependencies": {
-    "@types/vscode": "^1.91.0",
+    "@types/vscode": "^1.93.0",
     "@types/mocha": "^10.0.7",
     "@types/node": "20.x",
-    "eslint": "^8.57.0",
-    "typescript": "^5.4.5",
-    "@vscode/test-cli": "^0.0.9",
-    "@vscode/test-electron": "^2.4.0"
+    "@typescript-eslint/eslint-plugin": "^8.3.0",
+    "@typescript-eslint/parser": "^8.3.0",
+    "eslint": "^9.9.1",
+    "typescript": "^5.5.4",
+    "@vscode/test-cli": "^0.0.10",
+    "@vscode/test-electron": "^2.4.1"
   }
 }
