@@ -1,9 +1,9 @@
-const assert = require('assert');
+import * as assert from 'assert';
 
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
-const vscode = require('vscode');
-// const myExtension = require('../extension');
+import * as vscode from 'vscode';
+// import * as myExtension from '../../extension';
 
 suite('Extension Test Suite', () => {
 	vscode.window.showInformationMessage('Start all tests.');
